+++ scripts/generators/index.js
@@ -2,6 +2,7 @@
 
 "use strict";
 
+module.exports = (hexo) => {
 hexo.extend.generator.register("doratiger_404", require("./lib/404"));
 hexo.extend.generator.register("doratiger_about", require("./lib/about"));
 hexo.extend.generator.register("doratiger_tags", require("./lib/tags"));
@@ -17,3 +18,4 @@ hexo.extend.generator.register("doratiger_terms", require("./lib/terms"));
 hexo.extend.generator.register("doratiger_privacy", require("./lib/privacy"));
 hexo.extend.generator.register("doratiger_sitemap", require("./lib/sitemap"));
 hexo.extend.generator.register("doratiger_local_search", require("./lib/local-search"));
+};